+++ src/kodo_python/coder.hpp
@@ -5,35 +5,36 @@
 
 #pragma once
 
-#include <string>
+#include <pybind11/pybind11.h>
+#include <pybind11/functional.h>
 
-#include <boost/python.hpp>
+#include <string>
+#include <functional>
 
 namespace kodo_python
 {
 
-template<class Coder>
-void set_trace_callback(Coder& coder, PyObject* function)
-{
-    auto callback = [function](
-        const std::string& zone, const std::string& message)
-    {
-        boost::python::call<void>(function, zone, message);
-    };
+/// The callback used to output the logged information
+using log_callback =
+    std::function<void(const std::string& zone, const std::string& data)>;
 
+template<class Coder>
+void set_trace_callback(Coder& coder, const log_callback& callback)
+{
     coder.set_trace_callback(callback);
 }
 
 template<class Coder>
-auto coder(const std::string& name) ->
-    boost::python::class_<Coder, boost::noncopyable>
+pybind11::class_<Coder> coder(pybind11::module& m, const std::string& name)
 {
-    using namespace boost::python;
+    using namespace pybind11;
 
     using coder_type = Coder;
 
-    auto coder_class = class_<coder_type, boost::noncopyable>(
-        name.c_str(), "An (en/de)coder", no_init)
+    // Declare std::shared_ptr as a holder type, because the factories
+    // return a shared_ptr when building a coder instance
+    auto coder_class = class_<coder_type, std::shared_ptr<coder_type>>(
+        m, name.c_str(), "A coder object")
         .def("payload_size", &coder_type::payload_size,
              "Return the required payload buffer size in bytes.\n\n"
              "\t:returns: The required payload buffer size in bytes.\n")
@@ -48,7 +49,7 @@ auto coder(const std::string& name) ->
              "Return the number of symbols in this block coder.\n\n"
              "\t:returns: The number of symbols in this block coder.\n")
         .def("set_trace_callback", &set_trace_callback<coder_type>,
-             boost::python::arg("callback"),
+             arg("callback"),
              "Write the trace information to a callback.\n\n"
              "\t:param callback: The callback which is called with the zone "
              "and message.")
@@ -57,11 +58,11 @@ auto coder(const std::string& name) ->
         .def("set_trace_off", &coder_type::set_trace_off,
              "Disable tracing.\n")
         .def("set_zone_prefix", &coder_type::set_zone_prefix,
-             boost::python::arg("zone_prefix"),
+             arg("zone_prefix"),
              "Set a zone prefix for the tracing output.\n\n"
              "\t:param zone_prefix: The zone prefix to append to all "
              "tracing zones.");
 
-    return coder_class;
+    return std::move(coder_class);
 }
 }