+++ src/kodo_python/symbol_decoding_status_updater_methods.hpp
@@ -12,17 +12,17 @@ static void symbol_decoding_status_updater_methods(DecoderClass& decoder_class)
 {
     decoder_class
     .def("set_status_updater_on",
-         &DecoderClass::wrapped_type::set_status_updater_on,
+         &DecoderClass::type::set_status_updater_on,
          "Sets the status updater on.\n\n")
     .def("set_status_updater_off",
-         &DecoderClass::wrapped_type::set_status_updater_off,
+         &DecoderClass::type::set_status_updater_off,
          "Sets the status updater off.\n\n")
     .def("update_symbol_status",
-         &DecoderClass::wrapped_type::update_symbol_status,
+         &DecoderClass::type::update_symbol_status,
          "Updates the symbol status so that all uncoded symbols, "
          "label partially complete.\n\n")
     .def("is_status_updater_enabled",
-         &DecoderClass::wrapped_type::is_status_updater_enabled,
+         &DecoderClass::type::is_status_updater_enabled,
          "Check if the symbol status updater is enabled or not.\n\n"
          "\t:returns: True if enabled, otherwise false.\n");
 }